feature: Movement w.r.t View direction 1. camera will move with respect to the current camera view direction Note: need bug fixing for UP and DOWN movement after performing yaw rotation once

+++ geom/geom.go
@@ -3,10 +3,10 @@ package geom
 import "math"
 
 // pi radian = 180 deg
-func DegreeToRadian(degree float32) float32 {
+func DegreeToRadian(degree float64) float64 {
 	return degree * (math.Pi / 180)
 }
 
-func RadianToDegree(radian float32) float32 {
+func RadianToDegree(radian float64) float64 {
 	return radian * (180 / math.Pi)
 }
